// airbnb/interfaces/listing.geo.ts
export interface GeolistingData {
  type: string;
  geometry: {
    type: string;
    coordinates: [number, number];
  };
  properties: {
    id: string;
    listing_url: string;
    scrape_id: string;
    last_scraped: string;
    name: string;
    summary: string;
    space: null | string;
    description: string;
    experiences_offered: string;
    neighborhood_overview: null | string;
    notes: null | string;
    transit: null | string;
    access: null | string;
    interaction: null | string;
    house_rules: string;
    thumbnail_url: string;
    medium_url: string;
    picture_url: {
      thumbnail: boolean;
      filename: string;
      format: string;
      width: number;
      mimetype: string;
      etag: string;
      id: string;
      last_synchronized: string;
      color_summary: string[];
      height: number;
    };
    xl_picture_url: string;
    host_id: string;
    host_url: string;
    host_name: string;
    host_since: string;
    host_location: string;
    host_about: null | string;
    host_response_time: string;
    host_response_rate: number;
    host_acceptance_rate: null | string;
    host_thumbnail_url: string;
    host_picture_url: string;
    host_neighbourhood: null | string;
    host_listings_count: number;
    host_total_listings_count: number;
    host_verifications: string[];
    street: string;
    neighbourhood: null | string;
    neighbourhood_cleansed: string;
    neighbourhood_group_cleansed: null | string;
    city: string;
    state: string;
    zipcode: string;
    market: string;
    smart_location: string;
    country_code: string;
    country: string;
    latitude: string;
    longitude: string;
    property_type: string;
    room_type: string;
    accommodates: number;
    bathrooms: number;
    bedrooms: number;
    beds: number;
    bed_type: string;
    amenities: string[];
    square_feet: null | number;
    price: number;
    weekly_price: null | number;
    monthly_price: null | number;
    security_deposit: number;
    cleaning_fee: null | number;
    guests_included: number;
    extra_people: number;
    minimum_nights: number;
    maximum_nights: number;
    calendar_updated: string;
    has_availability: null | string;
    availability_30: number;
    availability_60: number;
    availability_90: number;
    availability_365: number;
    calendar_last_scraped: string;
    number_of_reviews: number;
    first_review: string;
    last_review: string;
    review_scores_rating: number;
    review_scores_accuracy: number;
    review_scores_cleanliness: number;
    review_scores_checkin: number;
    review_scores_communication: number;
    review_scores_location: number;
    review_scores_value: number;
    license: null | string;
    jurisdiction_names: null | string;
    cancellation_policy: string;
    calculated_host_listings_count: number;
    reviews_per_month: number;
    features: string[];
  };
}